Re-arrange how we download the server
Change the order of processing so wlpcfg is "built" first, and
it takes care of finding/stashing away a runtime
--> some work still to do here, as I can't get some of the remote
    options to work whatever I try.

wlpcfg generates/updates a properties file that indicates what the
wlp.install.dir is, and whether or not it is valid.

default is to continue if the wlp install can't be found, and
to use a profile when the build should fail if all tests can't
be run.